+++ ui/pages/index.vue
@@ -1,13 +1,27 @@
 <script setup lang="ts">
-import TypeLink from "~/components/helpers/TypeLink.vue";
+import {FontAwesomeIcon} from "@fortawesome/vue-fontawesome";
 </script>
 
 <template>
-	<div>
-		<h1>Index Page</h1>
-		<ul>
-			<li><NuxtLink to="/categories">Categories</NuxtLink></li>
-			<li><TypeLink type-id="645"></TypeLink></li>
-		</ul>
+	<div class="mt-4 grid grid-cols-3 gap-4">
+
+		<div class="border-2 p-4">
+			<h2>New to EVE Online?</h2>
+			<p>Get 1,000,000 free skill points.</p>
+			<a href="https://www.eveonline.com/signup?invc=b28d194d-7181-4bf0-8e3f-72cebbc7ca7d">Join New Eden</a>
+		</div>
+
+		<div class="border-2 p-4">
+			<h2>Already Playing?</h2>
+			<p>Use code "everef" at Markee Dragon to get 3% off PLEX and Omega.</p>
+			<a href="https://store.markeedragon.com/affiliate.php?id=933">Get PLEX</a>
+		</div>
+
+		<div class="border-2 p-4">
+			<h2>Love EVE Ref?</h2>
+			<p>Support the project on Patreon.</p>
+			<a href="https://patreon.com/everef">Donate</a>
+		</div>
+
 	</div>
 </template>