+++ ui/plugins/fontawesome.js
@@ -1,0 +1,14 @@
+import { library, config } from "@fortawesome/fontawesome-svg-core";
+import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
+import { faPatreon, faGithub, faDiscord } from "@fortawesome/free-brands-svg-icons";
+
+// This is important, we are going to let Nuxt worry about the CSS
+config.autoAddCss = false;
+
+// You can add your icons directly in this plugin. See other examples for how you
+// can add other styles or just individual icons.
+library.add(faPatreon, faGithub, faDiscord);
+
+export default defineNuxtPlugin((nuxtApp) => {
+	nuxtApp.vueApp.component("font-awesome-icon", FontAwesomeIcon, {});
+});